requirements.txt
coverage==5.3
defusedcsv~=1.1.0
defusedxml~=0.6.0
Django~=3.2.4
django-allauth~=0.44.0
django-cleanup~=5.1.0
django-compressor==2.4
django-extensions~=3.1.0
django-filter~=2.4.0
django-libsass==0.8
django-mathfilters~=1.0.0
django-mptt~=0.11.0
django-rest-swagger~=2.2.0
django-settings-export~=1.2.1
django-widget-tweaks~=1.4.8
djangorestframework~=3.12.2
drf-extensions~=0.7.0
iso8601~=0.1.14
jsonfield~=3.0.0
Markdown~=3.2.2
pypandoc==1.5
pytest~=6.1.2
pytest-cov~=2.11.1
pytest-django~=4.1.0
pytest-dotenv~=0.5.2
pytest-mock~=3.5.1
pytest-pythonpath~=0.7.3
rules==2.2
